+++ src/partials/work_together.html
@@ -1,6 +1,6 @@
 <!-- html код для секції work together -->
 <footer id="footer-id">
-  <div class="container">
+  <div class="container footer-container">
     <div class="title-box">
       <picture>
         <source media="(min-width:768px)" srcset="
@@ -32,7 +32,7 @@
       </li>
     </ul>
   </div>
-  <div class="modal-background visually-hidden">
+  <div class="modal-background">
     <div class="modal container">
       <h3 class="modal-title">Thank you for your interest in cooperation!</h3>
       <p class="modal-text">
